feat: support GCP as a cloud provider
Implement bucket methods and operations to fully support cassandrabackup in GCP.
The only logical difference in the objects stored in GCP is that we can support
the reset of the object retention expiration (via EventBaseHolt). This allows us
to have automatic prunning by creating a bucket policy.

The blobs digests for GCP are the same as AWS (BLAKE2b hash algorithm) but the
logic has been extrated to a separate package. Additionally In GCP we don't need
to support multipart uploads for now, therefore we don't compute SHA256 and MD5
needed exclusivelly for AWS multipart.

Signed-off-by: Ivan Daunis <ivan.daunis@retailnext.net>