feat: support GCP as a cloud provider
Implement bucket methods and operations to fully support cassandrabackup in GCP.
The only logical difference in the objects stored in GCP is that we can support
the reset of the object retention expiration (via EventBaseHolt). This allows us
to have automatic prunning by creating a bucket policy.

The blobs digests for GCP are the same as AWS (BLAKE2b hash algorithm) the only
difference is for GCP we calculate an MD5 and the length of the entire object
instead of using the parts.PartDigests which computes SHA256 and MD5 for every part.

This change is backwards compatible, and existing digests cache are reusable for AWS.
A node can technically backup in both clouds without changing the cache
configuration.

Signed-off-by: Ivan Daunis <ivan.daunis@retailnext.net>